test: share a task_with_status scenario across mixed-enum tests
`mixed_enum_schema` and `mixed_enum_roundtrip` each redefined the same
mixed unit/data `Status { Pending, Failed, Done }` enum and a wrapper
model. Extract them into a `task_with_status` scenario; both tests now
run in both id modes.
