+++ crates/toasty-driver-integration-suite/src/scenarios.rs
@@ -16,6 +16,7 @@ pub mod has_many_unique_title;
 pub mod has_one_optional_belongs_to;
 pub mod in_list_item;
 pub mod tagged_item;
+pub mod task_with_status;
 pub mod two_models;
 pub mod two_models_unique_title;
 pub mod user_account_subscription;
